+++ compiler/gen/gen.v
@@ -260,7 +260,7 @@ fn (mut g Gen) typ(typ ast.Type) string {
 			g.typedefs.writeln('typedef struct $name $name;')
 			g.structs.writeln('struct $name {')
 			for i, f in (ts.info as ast.StructInfo).fields {
-				g.structs.writeln('   ${g.typ(f)} f${i+1};')
+				g.structs.writeln('   ${g.typ(f)} f${i + 1};')
 			}
 			g.structs.writeln('};\n')
 			g.types << idx